+++ test/integration/cookbooks/resolver-test/recipes/default.rb
@@ -1,4 +1,9 @@
-node.default['resolver']['search'] = ['localdomain']
-node.default['resolver']['domain'] = nil
-node.default['resolver']['nameservers'] = %w(1.1.1.1 8.8.8.8)
-node.default['resolver']['options'] = { 'timeout' => 2, 'rotate' => nil }
+resolver_config '/etc/resolv.conf' do
+  nameservers ['1.1.1.1', '1.0.0.1']
+  domain 'test.com'
+  search ['test1.com', 'test2.com']
+  options(
+    'timeout' => 2
+  )
+  override_system_configuration true
+end